% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-schedule.R
\name{get_staffing}
\alias{get_staffing}
\title{Gets the current staffing schedule from the server}
\usage{
get_staffing()
}
\value{
A dataframe with columns kerberos, shift, day, and time
}
\description{
Gets the current staffing schedule from the server
}
