% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{group_schedule}
\alias{group_schedule}
\title{Groups unprocessed dataframe into sessions}
\usage{
group_schedule(hours)
}
\arguments{
\item{hours}{A dataframe with kerberos, shift_day, shift_hour columns}
}
\value{
A dataframe with kerberos, session_day, start, end columns
  corresponding to the same staffing schedule as `hours`.
}
\description{
Groups unprocessed dataframe into sessions
}
