% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule-plots.R
\name{schedule_plot_attendance}
\alias{schedule_plot_attendance}
\title{Attendance plot}
\usage{
schedule_plot_attendance(schedule, pool = get_pool(), now_line = TRUE)
}
\arguments{
\item{schedule}{Lab schedule}

\item{pool}{database pool}

\item{now_line}{plot a red line for now?}
}
\value{
Nothing, plots
}
\description{
Attendance plot
}
