% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule-plots.R
\name{schedule_plot_vertical}
\alias{schedule_plot_vertical}
\title{Title}
\usage{
schedule_plot_vertical(schedule)
}
\arguments{
\item{schedule}{Lab schedule}
}
\value{
Nothing, plots
}
\description{
Title
}
