% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{shift_end_time}
\alias{shift_end_time}
\title{Get shift end time}
\usage{
shift_end_time(shifts)
}
\arguments{
\item{shifts}{vector of shifts ('Sat1700', 'Tue0900', ...)}
}
\value{
datetimes corresponding to the end of the shift.
}
\description{
Get shift end time
}
