% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{shift_start_time}
\alias{shift_start_time}
\alias{shift_to_time}
\title{Convert shift strings to datetimes}
\usage{
shift_start_time(shifts)
}
\arguments{
\item{shifts}{vector of shifts ('Sat1700', 'Tue0900', ...)}
}
\value{
datetimes corresponding to the next such shift (if the day is today,
  returns today's shift)
}
\description{
Convert shift strings to datetimes
}
