% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-queue.R
\name{staff_q}
\alias{staff_q}
\title{Get staff queue}
\usage{
staff_q(pool, start = floor_date(now(), unit = "day"), end = now())
}
\arguments{
\item{pool}{Connection to the database}

\item{start}{Start time}

\item{end}{End time}
}
\value{
A dataframe describing the staff side of the queue. Redundant rows.
}
\description{
Get staff queue
}
