% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-queue.R
\name{staff_queue_deltas}
\alias{staff_queue_deltas}
\title{Simplify staff queue}
\usage{
staff_queue_deltas(staff_queue_full)
}
\arguments{
\item{staff_queue_full}{A dataframe, output from staff_q()}
}
\description{
Simplify staff queue
}
