% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{staffing_by_shift}
\alias{staffing_by_shift}
\title{Who's on duty now?}
\usage{
staffing_by_shift(staffing)
}
\arguments{
\item{staffing}{Staff schedule (ungrouped)}
}
\value{
A nested dataframe sorted by shift, with two cols: `shift`,
  `staff`.Each row of `staff` contains a char vector of the staff on duty
  then. If no staff are on duty, staff is NA. Staff list
  sorted alphabetically.
}
\description{
Who's on duty now?
}
