% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-queue.R
\name{student_q}
\alias{student_q}
\title{Get student queue}
\usage{
student_q(pool, start = floor_date(now(), unit = "day"), end = now())
}
\arguments{
\item{pool}{Connection to the database}

\item{start}{Start time}

\item{end}{End time}
}
\value{
A dataframe describing the student side of the queue. Many redundant
  rows
}
\description{
Get student queue
}
