% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-queue.R
\name{student_queue_deltas}
\alias{student_queue_deltas}
\title{Simplify the student queue data}
\usage{
student_queue_deltas(student_q_full)
}
\arguments{
\item{student_q_full}{A dataframe, output from student_q()}
}
\value{
A simplified dataframe
}
\description{
Simplify the student queue data
}
