% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{time_to_gcal}
\alias{time_to_gcal}
\title{Format time for Google Calendar use}
\usage{
time_to_gcal(times)
}
\arguments{
\item{times}{times to convert}
}
\value{
Character vector of dates properly formatted
}
\description{
Format time for Google Calendar use
}
