% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{time_to_hour}
\alias{time_to_hour}
\title{Converts a date to a decimal hour component}
\usage{
time_to_hour(times)
}
\arguments{
\item{times}{datetimes to convert}
}
\value{
hours since the start of day (can be fractional)
}
\description{
Converts a date to a decimal hour component
}
