% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{time_to_shift}
\alias{time_to_shift}
\title{Convert datetime to shift string}
\usage{
time_to_shift(datetimes = now())
}
\arguments{
\item{datetimes}{vector of datetimes to convert}
}
\value{
vector of strings corresponding to the shift in which the time
  happens
}
\description{
Convert datetime to shift string
}
