% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-utils.R
\name{time_to_weekday}
\alias{time_to_weekday}
\title{Day string to ordered factor}
\usage{
time_to_weekday(times)
}
\arguments{
\item{times}{vector of strings to convert}
}
\value{
ordered factors for a week starting on Monday
}
\description{
Day string to ordered factor
}
