% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{staff_on_duty}
\alias{staff_on_duty}
\title{Who's on duty now?}
\usage{
staff_on_duty(staffing, shifts)
}
\arguments{
\item{staffing}{Staff schedule (ungrouped)}

\item{shifts}{What shifts are we interested in (not times)}
}
\value{
A nested dataframe, with two cols: `shift`, `staff`. Each row of
  `staff` contains a char vector of the staff on duty then.
}
\description{
Who's on duty now?
}
