% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-queue.R
\name{staff_q}
\alias{staff_q}
\title{Get staff queue}
\usage{
staff_q(pool)
}
\arguments{
\item{pool}{Connection to the database}
}
\value{
A dataframe describing the staff side of the queue. Redundant rows.
}
\description{
Get staff queue
}
